+++ decidim-proposals/app/helpers/decidim/proposals/application_helper.rb
@@ -95,6 +95,10 @@ module Decidim
         proposal.endorsements.for_listing.map { |identity| present(identity.normalized_author) }
       end
 
+      def form_has_address?
+        @form.address.present? || @form.has_address
+      end
+
       def authors_for(collaborative_draft)
         collaborative_draft.identities.map { |identity| present(identity) }
       end